Ticket 13: reset the schema to one account and one read state
Room restarts at version 1 with the entities of docs/article-store.md and no
migration. The article table is Article, keyed by the 64-bit id FreshRSS gives
it, with read, starred and read_at as columns of the row and a PendingChange
table for what the server has not been told yet. ItemState, ItemStateChange,
Tag, TagJoin, AccountType, AccountConfig.useSeparateState, the four migrations
and the six exported schemas are gone, along with every account-scoped join:
there is one account, in one row, and no other table carries an account id.

The account layer is a FreshRSS login screen. The account selection screen and
its dialog are deleted, logging in replaces the one account row instead of
inserting a second, and the account tab lost add, delete and "Other accounts".
Notifications keep working: the sync worker no longer carries an account id,
and its mark-read and star actions write through the repository as before.

An article's identity is parsed in one place, ArticleIds in the api module: the
hexadecimal long form of stream/contents and the decimal of stream/items/ids
become the same number, and the decimal is what goes back to the server.

The timeline query names its own join order, driving from the article table, so
the index that serves the ordering is the only sensible plan whether the
planner has statistics or not. On a hundred thousand seeded articles the first
page of every timeline takes 0.4 ms and the two drawer counts 1.1 ms, where
ticket 11 measured 90 ms, 73 ms and 26 ms on a hundred and ten thousand — and
all of them are now flat in the number of articles stored rather than linear in
it. TimelineTimeBudgetTest holds those budgets on the gate, before and after
PRAGMA optimize, because statistics turned out to make the drawer's per-feed
count sixty times slower until the query named the index it walks.

The sync keeps its inherited sequence with the least change that works against
the new store: it upserts articles by id, overwriting content columns only for
an id already held, writes state from the three id lists, and clears the queue
at the end. One transaction, batching and paging are ticket 14's; deleting old
articles is 15's; the history screen is 16's.

+++ app/src/main/java/app/lenews/item/ItemScreen.kt
@@ -33,7 +33,7 @@ import kotlinx.coroutines.flow.distinctUntilChanged
 import org.koin.core.parameter.parametersOf
 
 class ItemScreen(
-    private val itemId: Int,
+    private val itemId: Long,
     private val itemIndex: Int,
     private val queryFilters: QueryFilters
 ) : AndroidScreen() {
